OBPIH-6536 Move generating invoices to prepayment service
