+++ grails-app/controllers/org/pih/warehouse/invoice/InvoiceController.groovy
@@ -25,6 +25,7 @@ class InvoiceController {
     InvoiceDataService invoiceDataService
     InvoiceService invoiceService
     DocumentService documentService
+    PrepaymentInvoiceService prepaymentInvoiceService
 
     // This template is generated by webpack during application start
     def index() {
@@ -96,7 +97,7 @@ class InvoiceController {
             redirect(action: "list")
         }
 
-        Invoice invoice = invoiceService.generatePrepaymentInvoice(order)
+        Invoice invoice = prepaymentInvoiceService.generatePrepaymentInvoice(order)
         redirect(action: "create", params: [id: invoice.id])
     }
 
@@ -107,7 +108,7 @@ class InvoiceController {
             redirect(action: "list")
         }
 
-        Invoice invoice = invoiceService.generateInvoice(order)
+        Invoice invoice = prepaymentInvoiceService.generateInvoice(order)
         redirect(action: "create", params: [id: invoice.id])
     }
